Split enterprise and community builds
