+++ src/app/dynamic/common.ts
@@ -1,0 +1,5 @@
+import {environment} from '../../environments/environment';
+
+export function getDynamicModulesRootPath() {
+  return `src/app/dynamic/${environment.edition}`;
+}
