+++ src/app/dynamic/enterprise/README.md
@@ -1,0 +1,2 @@
+## Important
+Contents of this directory are available only in Enterprise Edition.
