+++ src/app/dynamic/enterprise/theming/picker/style.scss
