+++ src/app/dynamic/enterprise/theming/services/theme.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
-import {AppConfigService} from '../../app-config.service';
-import {Theme} from '../../shared/model/Config';
+import {Theme} from '../../../../shared/model/Config';
+import {AppConfigService} from '../../../../app-config.service';
 
 @Injectable()
 export class ThemeService {
