+++ src/app/settings/user/user-settings.routing.module.ts
@@ -4,6 +4,7 @@ import {RouterModule, Routes} from '@angular/router';
 import {AuthGuard} from '../../core/services';
 
 import {UserSettingsComponent} from './user-settings.component';
+import {getDynamicModulesRootPath} from '../../dynamic/common';
 
 const routes: Routes = [
   {
@@ -13,7 +14,8 @@ const routes: Routes = [
     children: [
       {
         path: '',
-        loadChildren: () => import('../../theming/module').then(module => module.ThemingModule),
+        loadChildren: () =>
+          import(`${getDynamicModulesRootPath()}/theming/module`).then(module => module.ThemingModule),
         data: {preload: true},
       },
     ],
